Changed vect to be composed of floats, not ints
This meant changing the typedef to Vector2f instead of Vector2i. This
also meant that I had to start using the 'round' function and then
casting the result as an int when using the position as an array
subscript.

+++ vector_misc.hpp
@@ -3,7 +3,7 @@
 #define VECTOR_MISC_HPP
 
 // Convenience type for vectors
-typedef sf::Vector2i vect;
+typedef sf::Vector2f vect;
 
 std::string velocity_to_str(vect v);
 std::string position_to_str(vect v);
